+++ vite-project/component/Header.js
@@ -9,6 +9,7 @@ export function Header(item) {
     const icon = document.createElement('a')
     const img = document.createElement('img')
     
+    const register_link = document.createElement('a')
     
 head.classList.add("head")
 nav_menu.classList.add("nav_menu")
@@ -18,16 +19,20 @@ main.href = ""
 wallets.href = ""
 tranzactions.href = ""
 icon.href = ""
+register_link.href = "pages/signup/idnex.html"
 
 img.src = "../public/icons/log-out (1) 1.png"
+
 
 main.innerText = "Главная"
 wallets.innerText = "Мои кошельки"
 tranzactions.innerText = "Мои транзакции"
 h2.innerText = item.email
 
-head.append(nav_menu,user_data)
-nav_menu.append(main,wallets,tranzactions)
+register_link.innerText = "register"
+
+head.append(nav_menu,user_data,)
+nav_menu.append(main,wallets,tranzactions,register_link)
 user_data.append(h2,icon)
 icon.append(img)
