+++ vite-project/index.html
@@ -10,6 +10,7 @@
   <section class="header">
     <div class="container">
        
+      <a href="/separate/vite-project/pages/signup/idnex.html">перенос на Регистрация</a>
     </div>
 </section>
   <script type="module" src="./index.js"></script>
