+++ vite-project/pages/signin/index.html
@@ -9,20 +9,18 @@
 <body>
     
     <div class="register_container">
-        <h2>Регистрация</h2>
+        <h2>Войти в аккаунт</h2>
         <br><br>
-        <form name="register">
+        <form name="signin">
           <input type="email" name="email" placeholder="Почта" >
           <br><br>
-          <input type="text" name="name" placeholder="Имя">
-          <br><br>
-          <input type="text" name="surname" placeholder="Фамилия">
-          <br><br>
+         
+          
           <input type="password" name="password" placeholder="Пароль">
           <br><br>
-          <button>Продолжить</button>
+          <button><a href="/separate/vite-project/index.html">Продолжить</a></button>
           <br><br>
-          <a href="../otherPages/header.html">Войти в аккаунт</a>
+          <a href="/separate/vite-project/pages/signup/idnex.html">Регистрация</a>
         </form>
       </div>
