+++ vite-project/pages/signup/idnex.html
@@ -20,9 +20,9 @@
           <br><br>
           <input type="password" name="password" placeholder="Пароль">
           <br><br>
-          <button>Продолжить</button>
+          <button ><a href="/separate/vite-project/index.html">Продолжить</a></button>
           <br><br>
-          <a href="../otherPages/header.html">Войти в аккаунт</a>
+          <a href="/separate/vite-project/pages/signin/index.html">Войти в аккаунт</a>
         </form>
       </div>
